openvino: split quantization layout helpers
